# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Documentation",
    "description": "API documentation for your application"
  },
  "host": "localhost:5001",
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
    "/auth": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate user and get token",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User authenticated and token returned"
          },
          "400": {
            "description": "Invalid credentials or account not verified"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/send-otp": {
      "post": {
        "tags": ["Users"],
        "summary": "Send OTP to user email",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP sent to email"
          },
          "400": {
            "description": "User already exists"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/verify-otp": {
      "post": {
        "tags": ["Users"],
        "summary": "Verify OTP and create user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "otp": { "type": "string" }
              },
              "required": ["email", "otp"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User created and token returned"
          },
          "400": {
            "description": "Invalid OTP"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/delete": {
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" }
              },
              "required": ["email"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "400": {
            "description": "Invalid token"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/password/forgot": {
      "post": {
        "tags": ["Password"],
        "summary": "Request password reset OTP",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" }
              },
              "required": ["email"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP sent to email"
          },
          "400": {
            "description": "User does not exist"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/password/reset": {
      "post": {
        "tags": ["Password"],
        "summary": "Reset password using OTP",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "otp": { "type": "string" },
                "password": { "type": "string" }
              },
              "required": ["email", "otp", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset successful"
          },
          "400": {
            "description": "Invalid or expired OTP"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },

    "/coins/": {
      "get": {
        "tags": ["Coins"],
        "summary": "Get coins from user's account",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" }
              },
              "required": ["email"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coins Fetched successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/coins/add": {
      "post": {
        "tags": ["Coins"],
        "summary": "Add coins to user's account",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "coins": { "type": "number" }
              },
              "required": ["coins"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coins added successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/coins/use": {
      "put": {
        "tags": ["Coins"],
        "summary": "Use coins from user's account",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "coins": { "type": "number" }
              },
              "required": ["coins"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coins used successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/hints/": {
      "get": {
        "tags": ["Hints"],
        "summary": "Get hints from user's account",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" }
              },
              "required": ["email"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "hints Fetched successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/hints/add": {
      "post": {
        "tags": ["Hints"],
        "summary": "Add hints to user's account",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "hints": { "type": "number" }
              },
              "required": ["hints"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "hints added successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/hints/use": {
      "put": {
        "tags": ["Hints"],
        "summary": "Use hints from user's account",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "hints": { "type": "number" }
              },
              "required": ["hints"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "hints used successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/levels/set": {
      "put": {
        "tags": ["Levels"],
        "summary": "Set Level of the user",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "level": { "type": "number" }
              },
              "required": ["email", "level"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coins used successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },

    "/levels/words": {
      "get": {
        "tags": ["Levels"],
        "summary": "Get word for the user's level",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" }
              },
              "required": ["email"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Word Fetched successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },

    "/levels/public/words": {
      "get": {
        "tags": ["Levels"],
        "summary": "Get word for the level provided",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "number" }
              },
              "required": ["level"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Word Fetched successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/words/add": {
      "post": {
        "tags": ["Words"],
        "summary": "Creates words",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "words": { "type": "array" }
              },
              "required": ["words"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Words Added successfully"
          },
          "400": {
            "description": "Invalid words"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/words/daily/add": {
      "post": {
        "tags": ["Words"],
        "summary": "Creates daily words",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DailyWord",
              "type": "object",
              "properties": {
                "words": [{ "word": "string", "meaning": "string" }]
              },
              "required": ["words"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Words Added successfully"
          },
          "400": {
            "description": "Invalid words"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/words/daily": {
      "get": {
        "tags": ["Words"],
        "summary": "Get word for the day provided",
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "day": { "type": "string" }
              },
              "required": ["day"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Word Fetched successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Word not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/info": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user info",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "day": { "type": "string" }
              },
              "required": ["email"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Info Fetched successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/sync": {
      "post": {
        "tags": ["Users"],
        "summary": "Sync User's Local Data",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "coins": { "type": "number" },
                "level": { "type": "number" },
                "dailyChallenge": { "type": "array" }
              },
              "required": ["email", "coins", "level", "dailyChallenge"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data Synced successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/daily/update-daily-challenge": {
      "post": {
        "tags": ["Daily"],
        "summary": "Update Daily Challenge Stats",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "JWT token"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "dailyChallenge": { "type": "string" }
              },
              "required": ["email", "dailyChallenge"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily challenge updated successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },

  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Auth": {
      "type": "object",
      "properties": {
        "email": { "type": "string" },
        "password": { "type": "string" }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "email": { "type": "string" },
        "password": { "type": "string" }
      }
    },
    "VerifyOTP": {
      "type": "object",
      "properties": {
        "email": { "type": "string" },
        "otp": { "type": "string" }
      }
    },
    "DailyWord": {
      "type": "object",
      "properties": {
        "word": { "type": "string" },
        "meaning": { "type": "string" }
      }
    }
  }
}